Move the setFont function to a browser specific file. Move the cmap code into its own file. Move the data persistence code into its own file.
Add a data description file.
Changed fetched* to urlBase* to make the state of the font data more
obvious.
rename BackendService.prototype.parseHeader to
BackendService.prototype.parseDataHeader
Adjust some comments.
Removed an unreferenced variable.
